--- src/pages/store/Store.styles.ts
@@ -1,19 +1,0 @@
-import styled from 'styled-components'
-
-export const StoreWrapper = styled.article`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 50px 0;
-`
-export const StoreMapWrapper = styled.section`
-  margin-top: 20px;
-  width: 80%;
-  min-width: 800px;
-  height: 500px;
-
-  .map {
-    width: 100%;
-    height: 100%;
-  }
-`
